+++ common/src/main/java/com/walmartlabs/concord/common/Task.java
@@ -4,6 +4,4 @@ package com.walmartlabs.concord.common;
  * Marker interface for BPM tasks.
  */
 public interface Task {
-
-    String getKey();
 }