+++ runner/src/main/java/com/walmartlabs/concord/runner/engine/LocaleTask.java
@@ -5,13 +5,8 @@ import com.walmartlabs.concord.common.Task;
 import javax.inject.Named;
 import java.util.Locale;
 
-@Named
+@Named("locale")
 public class LocaleTask implements Task {
-
-    @Override
-    public String getKey() {
-        return "locale";
-    }
 
     public String[] countries() {
         return Locale.getISOCountries();