+++ runner/src/main/java/com/walmartlabs/concord/runner/engine/LoggingTask.java
@@ -5,13 +5,8 @@ import org.slf4j.LoggerFactory;
 
 import javax.inject.Named;
 
-@Named
+@Named("log")
 public class LoggingTask implements Task {
-
-    @Override
-    public String getKey() {
-        return "log";
-    }
 
     public void info(String log, String s) {
         LoggerFactory.getLogger(log).info(s);